refactor: 자료형들의 형변환
총 상금을 int로 했을때 저장범위가 넘어갈 경우를 생각하여
총 상금 계산시 long으로 형변환을 하였다.
마찬가지로 수익률 계산시 double로 형변환 하여
결과가 소수점 둘째짜리까지 나올수 있도록 하였다.(반올림 할 수 있도록)

+++ src/main/java/lotto/domain/Print.java
@@ -25,7 +25,7 @@ public class Print {
         }
     }
 
-    public void profitRate(float profit){
+    public void profitRate(double profit){
         System.out.println(String.format("총 수익률은 %.1f", profit) + "%입니다.");
     }
 }
